requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.4
SQLAlchemy==2.0.49
Flask-SQLAlchemy==3.1.1
paho-mqtt==2.1.0
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.12.1
Werkzeug==3.0.1
